backend/requirements.txt
flask
flask-cors
transformers
torch
seqeval
PyJWT
psycopg2-binary
python-dotenv
